Module Encodings.UCS

val min_value : int32

val max_value : int32

val is_non_character : int32 -> bool

Returns true if and only if the given value corresponds to a UCS * non-character. Such non-characters are forbidden for use in open * interchange of Unicode text data, and include the following: * 1. values from 0xFDD0 to 0xFDEF; and * 2. values 0xnFFFE and 0xnFFFF, where (0x0 <= n <= 0x10). * See the Unicode 5.0 Standard, section 16.7 for further details.

val is_out_of_range : int32 -> bool

Returns true if and only if the given value lies outside the * entire UCS range.

val is_surrogate : int32 -> bool

Returns true if and only if the given value corresponds to a UCS * surrogate code point, only for use in UTF-16 encoded strings. * See the Unicode 5.0 Standard, section 16.6 for further details.