Module Encodings.XML

val is_forbidden_control_character : int32 -> bool

Returns true if and only if the given value corresponds to * a forbidden control character as defined in section 2.2 of * the XML specification, version 1.0.