Module type Encodings.CHARACTER_DECODER

val decode_character : string -> int -> int32 * int

Decodes a single character embedded within a string. Given a string * and an index into that string, returns a tuple (value, width) where: * value = the value of the character at the given index; and * width = the width of the character at the given index, in bytes. * Raises an appropriate error if the character is invalid.