Module type Encodings.CHARACTER_ENCODER

val encode_character : int32 -> string

Encodes a single character value, returning a string containing * the character. Raises an error if the character value is invalid.