Module Xapi_stdext_encodings__Encodings.String_validator

Parameters

Decoder : CHARACTER_DECODER

Signature

val is_valid : string -> bool

Returns true if and only if the given string is validly-encoded.

val validate : string -> unit

Raises an encoding error if the given string is not validly-encoded.

val longest_valid_prefix : string -> string

Returns the longest validly-encoded prefix of the given string.