Module Encodings.XML

val is_illegal_control_character : Stdlib.Uchar.t -> bool

Returns true if and only if the given value corresponds to * a illegal control character as defined in section 2.2 of * the XML specification, version 1.0.